Move `Node` to a sealed interface and rework old `Node` class to `AbstractNode`
This is so that we can have more type specific interfaces like `Definition` inherit from `Node`, so that we don't need to fight around with the Kotlin type system by doing a bunch of casts.
